refactor: 패키지 이동 및 컨트롤러 네이밍 변경
