+++ app/src/main/java/com/techcourse/controller/ViewController.java
@@ -9,9 +9,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Controller
-public class AnnotationViewController {
+public class ViewController {
 
-    private static final Logger log = LoggerFactory.getLogger(AnnotationViewController.class);
+    private static final Logger log = LoggerFactory.getLogger(ViewController.class);
 
     @RequestMapping(value = "/", method = RequestMethod.GET)
     public String homeView(HttpServletRequest request, HttpServletResponse response) {
